updated checkout
