+++ client/src/components/ProductDetail/ProductDetailMain.jsx
@@ -3,7 +3,7 @@ import ProductInfo from '../../containers/productDetails/ProductInfoContainer.js
 import StyleSelector from '../../containers/productDetails/StyleSelectorContainer.js';
 import AddToCart from '../../containers/productDetails/AddToCartContainer.js';
 import ImageGallery from '../../containers/productDetails/ImageGalleryContainer.js';
-import ProductOverview from '../../containers/productDetails/productOverviewContainer.js';
+import ProductOverview from '../../containers/productDetails/ProductOverviewContainer.js';
 const ProductDetailMain = () => (
   <div className='product-display-container'>
     <div className='image-container'>
