+++ client/src/components/ProductDetail/ScrollButton.jsx
@@ -17,11 +17,10 @@ const ScrollButton = ({
         }}
       >
         <svg
-          // xmlns='http://www.w3.org/2000/svg'
+          xmlns='http://www.w3.org/2000/svg'
           width='25'
           height='25'
           fill='currentColor'
-          className='bi bi-arrow-left-short'
           viewBox='0 0 16 16'
         >
           <path
@@ -43,7 +42,6 @@ const ScrollButton = ({
           width='25'
           height='25'
           fill='currentColor'
-          className='bi bi-arrow-right-short'
           viewBox='0 0 16 16'
         >
           <path
